rustc_target: move in type definitions from rustc_trans::abi.
