+++ src/librustc_target/Cargo.toml
@@ -9,10 +9,10 @@ path = "lib.rs"
 crate-type = ["dylib"]
 
 [dependencies]
+bitflags = "1.0"
+log = "0.4"
 syntax = { path = "../libsyntax" }
 serialize = { path = "../libserialize" }
-log = "0.4"
-rand = "0.4"
 
 [features]
 jemalloc = []
