+++ src/librustc_target/abi/mod.rs
@@ -16,6 +16,8 @@ use spec::Target;
 use std::cmp;
 use std::ops::{Add, Sub, Mul, AddAssign, RangeInclusive};
 
+pub mod call;
+
 /// Parsed [Data layout](http://llvm.org/docs/LangRef.html#data-layout)
 /// for a target, which contains everything needed to compute layouts.
 pub struct TargetDataLayout {
