+++ src/librustc_trans/Cargo.toml
@@ -10,7 +10,6 @@ crate-type = ["dylib"]
 test = false
 
 [dependencies]
-bitflags = "1.0"
 cc = "1.0.1"
 flate2 = "1.0"
 jobserver = "0.1.5"
